+++ src/fileformats/c1cobfile.cpp
@@ -56,7 +56,7 @@ c1cobfile read_c1cobfile(std::istream& in) {
 
 	if (cob.picture.width > 0 && cob.picture.height > 0) {
 		cob.picture.data = shared_array<uint8_t>(cob.picture.width * cob.picture.height);
-		for (size_t i = 0; i < cob.picture.height; ++i) {
+		for (int32_t i = 0; i < cob.picture.height; ++i) {
 			in.read((char*)cob.picture.data.data() + cob.picture.width * (cob.picture.height - 1 - i), cob.picture.width);
 		}
 		cob.picture.colorkey = Color{0, 0, 0, 255};
